Address presumbit errors.
1. Remove ldap after TODO.
2. Move DISALLOW_COPY_AND_ASSIGN into private section.
3. Remove redundant blank lines.
4. Add explicit to ctors with single parameter and remove explicit for
   ctors with multiple parameters.
5. Add required STL header.
6. Replace DCHECK(a == b) with DCHECK_EQ(a, b) and do the same thing for
   similar checks.
7. Replace unsigned long long with uint64
8. Remove unnecessary ';' after '}'.
9. Add comment to end of namespace bracket.
10. Move static const into anonymous namespace.
11. Add {} to if statement with multiline body.
12. Replace C style cast with C++ style cast.

b/34275215

+++ cobalt/media/formats/webm/webm_cluster_parser_unittest.cc
@@ -270,7 +270,7 @@ bool VerifyTextBuffers(const scoped_ptr<WebMClusterParser>& parser,
 
 void VerifyEncryptedBuffer(scoped_refptr<StreamParserBuffer> buffer) {
   EXPECT_TRUE(buffer->decrypt_config());
-  EXPECT_EQ(static_cast<unsigned long>(DecryptConfig::kDecryptionKeySize),
+  EXPECT_EQ(static_cast<size_t>(DecryptConfig::kDecryptionKeySize),
             buffer->decrypt_config()->iv().length());
 }
